+++ src/main/kotlin/com/aoc/fuel/factory/components/Fuel.kt
@@ -2,7 +2,7 @@ package com.aoc.fuel.factory.components
 
 import java.util.*
 
-class Fuel(quantity: Double) : ReactionComponent("FUEL", quantity) {
+class Fuel(quantity: Int) : ReactionComponent("FUEL", quantity) {
     override fun equals(other: Any?): Boolean {
         if (other == null || other !is Fuel) return false
         return quantity == other.quantity
