+++ src/main/kotlin/com/aoc/fuel/factory/components/Ore.kt
@@ -2,7 +2,7 @@ package com.aoc.fuel.factory.components
 
 import java.util.*
 
-class Ore(quantity: Double) : ReactionComponent("ORE", quantity) {
+class Ore(quantity: Int) : ReactionComponent("ORE", quantity) {
     override fun equals(other: Any?): Boolean {
         if (other == null || other !is Ore) return false
         return quantity == other.quantity
